processors/cv/cv_processor.py
import os
import time
import json
import logging
from kafka import KafkaConsumer
from server.lib.tracing.trace_manager import traced
from server.lib.producers.enhanced_producer import EnhancedProducer
from server.lib.backpressure.backpressure_manager import BackpressureManager

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger("CVProcessor")

# Configuration
KAFKA_BOOTSTRAP_SERVERS = os.getenv("KAFKA_BOOTSTRAP_SERVERS", "kafka:29092")
INPUT_TOPIC = "osin.raw.satellite"
OUTPUT_TOPIC = "osin.processed.cv"
SERVICE_NAME = "cv-processor"

@traced(SERVICE_NAME, "analyze_imagery")
def run_imagery_analysis(event_id, _trace_context=None):
    """Execution span for simulated computer vision analysis"""
    logger.info(f"Analyzing imagery for Satellite Event: {event_id}")
    
    # Structural Stub: simulated detection results
    simulated_detections = [
        {"object": "infrastructure_anomaly", "confidence": 0.89, "box": [100, 200, 50, 50]},
        {"object": "thermal_anomaly", "confidence": 0.94, "box": [300, 150, 30, 30]}
    ]
    time.sleep(0.5) # Simulate processing time
    return simulated_detections

def main():
    # Initialize SVP-2 Enhanced Components
    producer = EnhancedProducer(KAFKA_BOOTSTRAP_SERVERS, SERVICE_NAME)
    backpressure = BackpressureManager(service_name=SERVICE_NAME)
    
    # Initialize Kafka Consumer
    consumer = None
    while not consumer:
        try:
            consumer = KafkaConsumer(
                INPUT_TOPIC,
                bootstrap_servers=KAFKA_BOOTSTRAP_SERVERS,
                value_deserializer=lambda v: json.loads(v.decode('utf-8')),
                group_id='osin-cv-group',
                auto_offset_reset='latest'
            )
            logger.info("Connected to Kafka Consumer")
        except Exception as e:
            logger.warning(f"Waiting for Kafka... {e}")
            time.sleep(5)

    logger.info(f"OSIN {SERVICE_NAME} Started with SVP-2 Enhancements")
    
    for message in consumer:
        try:
            # SVP-2 Backpressure Protection
            backpressure.check_and_throttle()
            
            data = message.value
            event_id = data.get('id')
            
            # SVP-1/2 Distributed Tracing: Extract parent context
            parent_trace = data.get('_trace')
            
            # Run Traced Imagery Analysis
            detections = run_imagery_analysis(event_id, _trace_context=parent_trace)
            
            # Enrich data
            processed_event = {
                "original_id": event_id,
                "source": data.get('source'),
                "processed_type": "cv_imagery_analysis",
                "intelligence": {
                    "detections": detections,
                    "resolution": "high",
                    "cloud_cover": 0.05
                },
                "timestamp": data.get('timestamp')
            }
            
            # Propagate Trace
            producer.send_traced(OUTPUT_TOPIC, processed_event, parent_context=parent_trace)
            logger.info(f"Processed CV Intelligence for: {event_id}")
                
        except Exception as e:
            logger.error(f"Error in CV processing: {e}")

if __name__ == "__main__":
    main()
