processors/cv/requirements.txt
kafka-python
opencv-python-headless
numpy
psutil
